refactor(Homepage): Move the grid page to the homepage
