--- pages/grid.tsx
@@ -1,10 +1,0 @@
-import React, { FC } from 'react'
-import Layout from '@components/Layout'
-import { HomeHeroSection } from '@components/HomeHeroSection'
-
-const Grid: FC = () => (
-	<Layout>
-		<HomeHeroSection />
-	</Layout>
-)
-export default Grid
